requirements.txt
fastapi == 0.111.0
localstack-client == 2.5
pandas == 2.2.2
requests == 2.32.2
uvicorn == 0.29.0
python-dotenv == 1.0.1
pytest == 8.2.1
boto3 == 1.34.113